+++ CHANGELOG.md
@@ -1,8 +1,11 @@
-# 1.0.0 (2020-08-15)
+# 1.0.0 (2020-08-16)
 
 
 ### Features
 
+* **build:** 新增 ESLint 代码校验能力 ([ccc5638](https://github.com/SUN-LG/typescript-starter/commit/ccc5638d97c20fd9a83e0b5808b623e893cde2dc))
+* **build:** 新增 Git Commit Message 提交规范及校验 ([d2ac9ac](https://github.com/SUN-LG/typescript-starter/commit/d2ac9acbfc99a3b7d06974f9b0ce744245fa8cbb))
+* **build:** 新增 Typescript 编译能力 ([0b0f05b](https://github.com/SUN-LG/typescript-starter/commit/0b0f05bfdc88ac806accda64099a831b58e0e90b))
 * 项目初始化 ([8b28b61](https://github.com/SUN-LG/typescript-starter/commit/8b28b61228e5f5e010c7a3388c9807bdcd2c97d4))